Compiled and edited

+++ src/SalesManager.java
@@ -14,18 +14,19 @@ public class SalesManager {
         }
         return max;
     }
-public int averageSales(){
+
+    public long averageSales() {
         int sum = 0;
-    for (int sale : sales) {
-        sum += sale;
-    }
-        return (sum - max()- min()) /(sales.length-2 );
+        for (long sale : sales) {
+            sum += sale;
+        }
+        return (sum - max() - min()) / (sales.length - 2);
     }
 
 
-    public int min() {
-        int min = Integer.MAX_VALUE;
-        for (int sale : sales) {
+    public long min() {
+        long min = Integer.MAX_VALUE;
+        for (long sale : sales) {
             if (sale < min) {
                 min = sale;
             }
